+++ CHANGELOG.md
@@ -10,6 +10,8 @@ ChangeLog
 - IMPROVEMENT: add Arch Linux platform support
 - IMPROVEMENT: add Fedora Linux platform support
 - IMPROVEMENT: add OpenSUSE Linux platform support
+- IMPROVEMENT: allow dotfiles to be added on command-line
+- IMPROVEMENT: allow ports to be mapped
 - BUGFIX: fix handling of execa() response object
 
 0.9.8 (2025-11-27)